Add fingerprinter.py
I wrote this script last night to generate fingerprints for a page's text content, based on the fingerprinting method described in last Thursday's lecture. I'm not sure how to deal with the calendar/history traps, so I thought that maybe we could use fingerprinting to detect near-duplicate pages and use that to help detect the traps. I don't know how efficient it will be, though.