Bug Fix
You may have fixed this locally, but whenever I ran the program an error was thrown. 
Line 75: compare_prints() takes in two arguments, but was passed three on line 75, so I removed the third argument. compare_prints() did not implement self.FINGERPRINT_OVERLAP_THRESHOLD so function is not effected and still works :)